Whitespace Changes + Bugfixes
I'm not sure why the whitespace wasn't proper before, but I'm doing a push thats just a few bug fixes alongside the whitespace change because the differences are terrible to look at.

Fixed a potential bug where a movement command could be sent without a point to move to.
-This was a confusing one, thanks to ArnaudB on Discord for the report and save.

The Barracks changes were not actually included the last version, because I'm a goof.
-After further testing, its cost is now a full 100k, and it now properly sticks to a count of 1 regardless of mark.