Movement Commands Bulked; Changed Build Process
Renamed XMLFiles to XMLMods, and updated the build to no longer directly insert itself into the game's folder, and instead mimic the game's folder structure within the project folder. This will allow the easy copying of both the dll and xml files in one easy copy and paste.

Added in new commands that will let us keep track of how often Civilians are moving around, and set movement commands to also be enacted in bulk.